MOved Album and Grade in Model package
